+++ PowerSync/PowerSync.Maui/CHANGELOG.md
@@ -1,8 +1,7 @@
 # PowerSync.Maui Changelog
 
 ## 0.0.4-alpha.1
-- Upstream PowerSync.Common version bump
-- Added ability to specify `AppMetadata` sync/stream requests (see Common changelog).
+- Upstream PowerSync.Common version bump (See Powersync.Common changelog for more information)
 
 ## 0.0.3-alpha.1
 - Upstream PowerSync.Common version bump