+++ PowerSync/PowerSync.Maui/CHANGELOG.md
@@ -1,11 +1,14 @@
 # PowerSync.Maui Changelog
 
+## 0.0.5-alpha.1
+- Upstream PowerSync.Common version bump (See Powersync.Common changelog 0.0.7-alpha.1 for more information)
+
 ## 0.0.4-alpha.1
-- Upstream PowerSync.Common version bump (See Powersync.Common changelog for more information)
+- Upstream PowerSync.Common version bump (See Powersync.Common changelog 0.0.6-alpha.1 for more information)
 - Added ability to specify `AppMetadata` sync/stream requests (see Common changelog).
 
 ## 0.0.3-alpha.1
-- Upstream PowerSync.Common version bump
+- Upstream PowerSync.Common version bump (See PowerSync.Common change 0.0.5-alpha.1 for more information)
 - Using the latest (0.4.9) version of the core extension, it introduces support for the Rust Sync implementation and also makes it the default - users can still opt out and use the legacy C# sync implementation as option when calling `connect()`.
 
 ## 0.0.2-alpha.1